requirements.txt
streamlit>=1.28.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
pypdf>=3.17.0
transformers>=4.45.0
bitsandbytes>=0.43.0
torch>=2.1.0
accelerate>=0.34.0
huggingface-hub>=0.24.0
PyMuPDF
pillow
opencv-python
safetensors
torchvision>=0.43.0